+++ packages/service-library/tests/test_redis_utils.py
@@ -1,8 +1,8 @@
 # pylint:disable=redefined-outer-name
 
 import asyncio
-from collections.abc import AsyncIterator
-from contextlib import asynccontextmanager
+from collections.abc import Callable
+from contextlib import AbstractAsyncContextManager
 from datetime import timedelta
 from itertools import chain
 from unittest.mock import Mock
@@ -14,7 +14,7 @@ from servicelib.background_task import stop_periodic_task
 from servicelib.redis import CouldNotAcquireLockError, RedisClientSDK
 from servicelib.redis_utils import exclusive, start_exclusive_periodic_task
 from servicelib.utils import logged_gather
-from settings_library.redis import RedisDatabase, RedisSettings
+from settings_library.redis import RedisDatabase
 from tenacity._asyncio import AsyncRetrying
 from tenacity.retry import retry_if_exception_type
 from tenacity.stop import stop_after_delay
@@ -23,22 +23,6 @@ from tenacity.wait import wait_fixed
 pytest_simcore_core_services_selection = [
     "redis",
 ]
-
-
-@asynccontextmanager
-async def get_redis_client_sdk(
-    redis_service: RedisSettings,
-) -> AsyncIterator[RedisClientSDK]:
-    redis_resources_dns = redis_service.build_redis_dsn(RedisDatabase.RESOURCES)
-    client = RedisClientSDK(redis_resources_dns)
-    assert client
-    assert client.redis_dsn == redis_resources_dns
-    await client.setup()
-
-    yield client
-    # cleanup, properly close the clients
-    await client.redis.flushall()
-    await client.shutdown()
 
 
 async def _is_locked(redis_client_sdk: RedisClientSDK, lock_name: str) -> bool:
@@ -52,9 +36,13 @@ def lock_name(faker: Faker) -> str:
 
 
 async def _contained_client(
-    redis_service: RedisSettings, lock_name: str, task_duration: float
+    get_redis_client_sdk: Callable[
+        [RedisDatabase], AbstractAsyncContextManager[RedisClientSDK]
+    ],
+    lock_name: str,
+    task_duration: float,
 ) -> None:
-    async with get_redis_client_sdk(redis_service) as redis_client_sdk:
+    async with get_redis_client_sdk(RedisDatabase.RESOURCES) as redis_client_sdk:
         assert not await _is_locked(redis_client_sdk, lock_name)
 
         @exclusive(redis_client_sdk, lock_key=lock_name)
@@ -70,20 +58,25 @@ async def _contained_client(
 
 @pytest.mark.parametrize("task_duration", [0.1, 1, 2])
 async def test_exclusive_sequentially(
-    redis_service: RedisSettings,
+    get_redis_client_sdk: Callable[
+        [RedisDatabase], AbstractAsyncContextManager[RedisClientSDK]
+    ],
     lock_name: str,
     task_duration: float,
 ):
-    await _contained_client(redis_service, lock_name, task_duration)
+    await _contained_client(get_redis_client_sdk, lock_name, task_duration)
 
 
 async def test_exclusive_parallel_lock_is_released_and_reacquired(
-    redis_service: RedisSettings, lock_name: str
+    get_redis_client_sdk: Callable[
+        [RedisDatabase], AbstractAsyncContextManager[RedisClientSDK]
+    ],
+    lock_name: str,
 ):
     parallel_tasks = 10
     results = await logged_gather(
         *[
-            _contained_client(redis_service, lock_name, task_duration=0.1)
+            _contained_client(get_redis_client_sdk, lock_name, task_duration=0.1)
             for _ in range(parallel_tasks)
         ],
         reraise=False
@@ -94,7 +87,7 @@ async def test_exclusive_parallel_lock_is_released_and_reacquired(
     ) == parallel_tasks - 1
 
     # check lock is being released
-    async with get_redis_client_sdk(redis_service) as redis_client_sdk:
+    async with get_redis_client_sdk(RedisDatabase.RESOURCES) as redis_client_sdk:
         assert not await _is_locked(redis_client_sdk, lock_name)
 
 
@@ -116,9 +109,12 @@ async def _assert_on_sleep_done(on_sleep_events: Mock, *, stop_after: float):
 
 
 async def _assert_task_completes_once(
-    redis_service: RedisSettings, *, stop_after: float
+    get_redis_client_sdk: Callable[
+        [RedisDatabase], AbstractAsyncContextManager[RedisClientSDK]
+    ],
+    stop_after: float,
 ) -> tuple[float, ...]:
-    async with get_redis_client_sdk(redis_service) as redis_client_sdk:
+    async with get_redis_client_sdk(RedisDatabase.RESOURCES) as redis_client_sdk:
         sleep_events = Mock()
 
         started_task = start_exclusive_periodic_task(
@@ -140,8 +136,12 @@ async def _assert_task_completes_once(
         return events_timestamps
 
 
-async def test_start_exclusive_periodic_task_single(redis_service: RedisSettings):
-    await _assert_task_completes_once(redis_service, stop_after=2)
+async def test_start_exclusive_periodic_task_single(
+    get_redis_client_sdk: Callable[
+        [RedisDatabase], AbstractAsyncContextManager[RedisClientSDK]
+    ]
+):
+    await _assert_task_completes_once(get_redis_client_sdk, stop_after=2)
 
 
 def _check_elements_lower(lst: list) -> bool:
@@ -158,12 +158,14 @@ def test__check_elements_lower():
 
 
 async def test_start_exclusive_periodic_task_parallel_all_finish(
-    redis_service: RedisSettings,
+    get_redis_client_sdk: Callable[
+        [RedisDatabase], AbstractAsyncContextManager[RedisClientSDK]
+    ]
 ):
     parallel_tasks = 10
     results: list[tuple[float, float]] = await logged_gather(
         *[
-            _assert_task_completes_once(redis_service, stop_after=60)
+            _assert_task_completes_once(get_redis_client_sdk, stop_after=60)
             for _ in range(parallel_tasks)
         ],
         reraise=False
